Separate IncrementalFont into a separate file. Cleanup the goog requires/provides.
